requirements.txt
prettytable
flask
twitter
beautifulsoup4
python-nvd3
unidecode
pymongo
textblob
nltk
unirest